Modify the file organization structure
