+++ opencompass/utils/__init__.py
@@ -1,6 +1,6 @@
 from .abbr import *  # noqa
 from .build import *  # noqa
-from .collect_env import *  # noqa
+# from .collect_env import *  # noqa
 from .fileio import *  # noqa
 from .git import *  # noqa
 from .lark import *  # noqa
